Made function name standard. Fixed some programs that didn't compile on linux. Refered some code that was copied from other places and had no reference. Moved python file to python folder Created folder to classify each type of algorithm Improved formatation of stack.c
