+++ C files/deletingMiddleElementInSingleyLinkedList.c
@@ -179,4 +179,3 @@ void displayList()
         }
     }
 }
-Output
