+++ C files/linkedListDeletion.c
