+++ C files/mergeTwoLnkedList.c
@@ -1,3 +1,10 @@
+/*
+ * mergeTwoLinkedList
+ *
+ * Solution for a challenge extracted from here:
+ * https://www.hackerrank.com/challenges/compare-two-linked-lists/problem
+ */
+
 #include <assert.h>
 #include <limits.h>
 #include <math.h>
