+++ C files/reverse_list.c
@@ -1,3 +1,10 @@
+/*
+ * reverse_list.c
+ *
+ * Solution from a challenge extracted from:
+ * https://www.hackerrank.com/challenges/print-the-elements-of-a-linked-list-in-reverse/problem
+ */
+
 #include <assert.h>
 #include <limits.h>
 #include <math.h>
