+++ Python files/delete_node.py
